+++ server/src/main/java/org/opensearch/action/bulk/TransportBulkAction.java
@@ -66,6 +66,7 @@ import org.opensearch.cluster.metadata.IndexNameExpressionResolver;
 import org.opensearch.cluster.metadata.MappingMetadata;
 import org.opensearch.cluster.metadata.Metadata;
 import org.opensearch.cluster.service.ClusterService;
+import org.opensearch.common.ValidationException;
 import org.opensearch.common.inject.Inject;
 import org.opensearch.common.lease.Releasable;
 import org.opensearch.common.unit.TimeValue;
@@ -540,12 +541,17 @@ public class TransportBulkAction extends HandledTransportAction<BulkRequest, Bul
                 if (docWriteRequest == null) {
                     continue;
                 }
+
                 if (addFailureIfRequiresAliasAndAliasIsMissing(docWriteRequest, i, metadata)) {
                     continue;
                 }
                 if (addFailureIfIndexIsUnavailable(docWriteRequest, i, concreteIndices, metadata)) {
                     continue;
                 }
+                if (addFailureIfAppendOnlyIndexAndOpsDeleteOrUpdate(docWriteRequest, i, concreteIndices, metadata)) {
+                    continue;
+                }
+
                 Index concreteIndex = concreteIndices.resolveIfAbsent(docWriteRequest);
                 try {
                     // The ConcreteIndices#resolveIfAbsent(...) method validates via IndexNameExpressionResolver whether
@@ -747,6 +753,47 @@ public class TransportBulkAction extends HandledTransportAction<BulkRequest, Bul
                     run();
                 }
             });
+        }
+
+        private boolean addFailureIfAppendOnlyIndexAndOpsDeleteOrUpdate(
+            DocWriteRequest<?> request,
+            int idx,
+            final ConcreteIndices concreteIndices,
+            Metadata metadata
+        ) {
+            Index concreteIndex = concreteIndices.resolveIfAbsent(request);
+            final IndexMetadata indexMetadata = metadata.index(concreteIndex);
+            if (indexMetadata.isAppendOnlyIndex()) {
+                if ((request.opType() == DocWriteRequest.OpType.UPDATE || request.opType() == DocWriteRequest.OpType.DELETE)) {
+                    ValidationException exception = new ValidationException();
+                    exception.addValidationError(
+                        "Operation ["
+                            + request.opType()
+                            + "] is not allowed as setting `"
+                            + IndexMetadata.INDEX_APPEND_ONLY_ENABLED_SETTING.getKey()
+                            + "` is enabled for this index: "
+                            + request.index()
+                    );
+                    addFailure(request, idx, exception);
+                    return true;
+                } else if (request.id() != null && request.opType() == DocWriteRequest.OpType.INDEX) {
+                    ValidationException exception = new ValidationException();
+                    exception.addValidationError(
+                        "Operation ["
+                            + request.opType()
+                            + "] is not allowed with a custom document id "
+                            + request.id()
+                            + " as setting `"
+                            + IndexMetadata.INDEX_APPEND_ONLY_ENABLED_SETTING.getKey()
+                            + "` is enabled for this index: "
+                            + request.index()
+                    );
+                    addFailure(request, idx, exception);
+                    return true;
+                }
+            }
+
+            return false;
         }
 
         private boolean addFailureIfRequiresAliasAndAliasIsMissing(DocWriteRequest<?> request, int idx, final Metadata metadata) {